+++ frontend/src/components/Index/Footer.vue
@@ -1,12 +1,14 @@
 <template>
     <footer>
         <div v-if="!core.footer" class="footer text-center mb-4 p-2">
-          <div class="d-block small text-dim">
-          <div class="d-block">
-            <router-link :to="$store.state.admin ? '/dashboard' : '/login'">{{$t('top_nav.dashboard')}}</router-link>
+          <div class="d-block text-dim">
+          <div class="mb-3">
+            <router-link :to="admin ? '/dashboard' : '/login'">{{$t('top_nav.dashboard')}}</router-link>
           </div>
-            <a class="mt-3" href="https://github.com/statping/statping" target="_blank">Statping</a> v{{core.version}}
-            made with <font-awesome-icon icon="heart" class="text-danger"/>
+            <span class="text-muted font-1 mt-3">
+            <a href="https://github.com/statping/statping" target="_blank">Statping v{{core.version}}
+              made with <font-awesome-icon icon="heart" class="font-1 text-danger"/></a>
+            </span>
           </div>
         </div>
         <div v-else class="footer text-center mb-4 p-2" v-html="core.footer"></div>
@@ -27,7 +29,10 @@
       },
     commit() {
       return this.$store.getters.core.commit.slice(0,8)
-    }
+    },
+    admin() {
+      return this.$store.getters.admin
+    },
   }
 }
 </script>