+++ frontend/src/components/Index/GroupServiceFailures.vue
@@ -1,17 +1,22 @@
 <template>
     <div>
-        <div class="d-flex mt-3 mb-2">
-            <div class="flex-fill service_day" v-for="(d, index) in failureData" :class="{'day-error': d.amount > 0, 'day-success': d.amount === 0}">
-                <span v-if="d.amount !== 0" class="d-none d-md-block text-center small">{{d.amount}}</span>
+        <div class="d-flex mt-3">
+            <div class="flex-fill service_day" v-for="(d, index) in failureData" @mouseover="mouseover(d)" @mouseout="mouseout" :class="{'day-error': d.amount > 0, 'day-success': d.amount === 0}">
+                <span v-if="d.amount !== 0" class="d-none d-md-block text-center small"></span>
             </div>
         </div>
         <div class="row mt-2">
-            <div class="col-3 text-left font-2 text-muted">30 Days Ago</div>
-            <div class="col-6 text-center font-2" :class="{'text-muted': service.online, 'text-danger': !service.online}">
-               {{service_txt}}
-            </div>
-            <div class="col-3 text-right font-2 text-muted">Today</div>
+          <div class="col-12 no-select">
+          <p class="divided">
+            <span class="font-2 text-muted">90 Days Ago</span>
+            <span class="divider"></span>
+            <span class="text-center font-2" :class="{'text-muted': service.online, 'text-danger': !service.online}">{{service_txt}}</span>
+            <span class="divider"></span>
+            <span class="font-2 text-muted">Today</span>
+          </p>
+          </div>
         </div>
+      <div class="daily-failures small text-right text-dim">{{hover_text}}</div>
     </div>
 </template>
 
@@ -26,6 +31,7 @@ export default {
     data() {
         return {
             failureData: [],
+          hover_text: ""
         }
     },
   props: {
@@ -43,18 +49,24 @@ export default {
       this.lastDaysFailures()
     },
     methods: {
+      mouseout() {
+        this.hover_text = ""
+      },
+    mouseover(e) {
+      let txt = `${e.amount} Failures`
+      if (e.amount === 0) {
+        txt = `No Issues`
+      }
+      this.hover_text = `${e.date.toLocaleDateString()} - ${txt}`
+    },
       async lastDaysFailures() {
-        const start = this.nowSubtract(86400 * 30)
-        const data = await Api.service_failures_data(this.service.id, this.toUnix(start), this.toUnix(this.startToday()), "24h")
+        const start = this.nowSubtract(86400 * 90)
+        const data = await Api.service_failures_data(this.service.id, this.toUnix(start), this.toUnix(this.now()), "24h", true)
         data.forEach((d) => {
           let date = this.parseISO(d.timeframe)
-          this.failureData.push({month: 1, day: date.getDate(), amount: d.amount})
+          this.failureData.push({month: date.getMonth(), day: date.getDate(), date: date, amount: d.amount})
         })
       }
     }
 }
 </script>
-
-<!-- Add "scoped" attribute to limit CSS to this component only -->
-<style scoped>
-</style>